Add support to retrieve TA load address with ASLR enabled
with the introduction of ASLR (Address Space Layout Randomization), it's no longer easy to find the correct place to set breakpoints for a Trusted Application.
With this script we can find automatically the correct load address to load symbols table, entering into "tee" and "ldelf" components of OP-TEE environment to retrieve the correct load address to use to
set breakpoints. When executing this script can connect to Normal Side UART to execute TA, then it sets automatically breakpoints on functions of tee and ldelf to retrieve TA load address.
Symbols table of executing TA is loaded and a breakpoint on TA_InvokeCommandEntryPoint is set accordingly.

Co-authored-by: Luca Gherardini <luca.gherardini96@gmail.com>
Co-authored-by: Simone Mione <simone.mione1@gmail.com>